requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
aiohttp==3.9.3 